Implement hold "sliding" samples in osu!mania
